feat: refactor a bunch of things into singleton hooks
* some fixes to make hot reload work better with singleton hooks
* add "showDeveloperModal" message handler
* add "setEnvironment" message handler
* add "setError" message handler
